+++ lib/controller.dart
@@ -4,6 +4,8 @@ import 'dart:math' as math;
 
 enum OrgChartOrientation { topToBottom, leftToRight }
 
+enum ActionOnNodeRemoval { unlink, connectToParent }
+
 class OrgChartController<E> {
   /// The list of nodes that we want to draw. this is generated from the items list.
   late List<Node<E>> _nodes;
@@ -13,7 +15,8 @@ class OrgChartController<E> {
   Size boxSize;
 
   /// The spacing between each node box. Needed here for the same reason i need boxSize here.
-  Offset spacing;
+  double spacing;
+  double runSpacing;
 
   /// The following function is included to ease the use of custom data types
   /// whether it be a map with key 'id' or 'pk' or a custom class, just use this function to provide the ids
@@ -25,19 +28,39 @@ class OrgChartController<E> {
   /// The function that returns the id of the node that the current node is pointing to.
   String? Function(E data) toProvider;
 
-  // setState Function
+  /// Used to change the to-id of the node when using removeItem function
+  /// if not provided, using removeItem will not be allowed!
+  void Function(E data, String? newID)? toSetter;
+
+  /// setState Function
   void Function(void Function() function)? setState;
 
-  OrgChartOrientation orientation;
+  OrgChartOrientation _orientation;
+  OrgChartOrientation get orientation => _orientation;
+  set orientation(OrgChartOrientation orientation) {
+    _orientation = orientation;
+    calculatePosition();
+  }
+
+  switchOrientation() {
+    _orientation = _orientation == OrgChartOrientation.topToBottom
+        ? OrgChartOrientation.leftToRight
+        : OrgChartOrientation.topToBottom;
+    calculatePosition();
+  }
 
   OrgChartController({
     required List<E> items,
     this.boxSize = const Size(200, 100),
-    this.spacing = const Offset(20, 50),
+    // this.spacing = const Offset(20, 50),
+    this.spacing = 20,
+    this.runSpacing = 50,
     required this.idProvider,
     required this.toProvider,
-    this.orientation = OrgChartOrientation.leftToRight,
-  }) : super() {
+    this.toSetter,
+    OrgChartOrientation orientation = OrgChartOrientation.leftToRight,
+  })  : _orientation = orientation,
+        super() {
     this.items = items;
   }
 
@@ -58,19 +81,19 @@ class OrgChartController<E> {
               i % 2 == 0
                   ? subNodes.length > i + 1 || subNodes.length == 1
                       ? 0
-                      : boxSize.width / 2 + spacing.dx / 2
-                  : spacing.dx + boxSize.width,
-              (_getLevel(subNodes[i]) + i ~/ 2) * (boxSize.height + spacing.dy),
+                      : boxSize.width / 2 + spacing / 2
+                  : spacing + boxSize.width,
+              (_getLevel(subNodes[i]) + i ~/ 2) * (boxSize.height + runSpacing),
             );
       }
       node.position = offset +
           Offset(
-            (subNodes.length > 1 ? boxSize.width / 2 + spacing.dx / 2 : 0),
-            _getLevel(node) * (boxSize.height + spacing.dy),
+            (subNodes.length > 1 ? boxSize.width / 2 + spacing / 2 : 0),
+            _getLevel(node) * (boxSize.height + runSpacing),
           );
       return (subNodes.length > 1
-          ? boxSize.width * 2 + spacing.dx * 3
-          : boxSize.width + spacing.dx * 2);
+          ? boxSize.width * 2 + spacing * 3
+          : boxSize.width + spacing * 2);
     } else {
       double dxOff = 0;
       for (var i = 0; i < subNodes.length; i++) {
@@ -85,12 +108,12 @@ class OrgChartController<E> {
       node.position = subNodes.length == 1
           ? Offset(
               subNodes.first.position.dx,
-              _getLevel(node) * (boxSize.height + spacing.dy),
+              _getLevel(node) * (boxSize.height + runSpacing),
             )
           : offset +
               Offset(
-                relOff / 2 - boxSize.width / 2 - spacing.dx,
-                _getLevel(node) * (boxSize.height + spacing.dy),
+                relOff / 2 - boxSize.width / 2 - spacing,
+                _getLevel(node) * (boxSize.height + runSpacing),
               );
       return relOff;
     }
@@ -106,25 +129,25 @@ class OrgChartController<E> {
       for (var i = 0; i < subNodes.length; i++) {
         subNodes[i].position = offset +
             Offset(
-              (_getLevel(subNodes[i]) + i ~/ 2) * (boxSize.width + spacing.dx),
+              (_getLevel(subNodes[i]) + i ~/ 2) * (boxSize.width + runSpacing),
               i % 2 == 0
                   ? subNodes.length > i + 1 || subNodes.length == 1
                       ? 0
-                      : boxSize.height / 2 + spacing.dy / 2
-                  : spacing.dy + boxSize.height,
+                      : boxSize.height / 2 + spacing / 2
+                  : spacing + boxSize.height,
             );
       }
       node.position = offset +
           Offset(
-            _getLevel(node) * (boxSize.width + spacing.dx),
-            (subNodes.length > 1 ? boxSize.height / 2 + spacing.dy / 2 : 0),
+            _getLevel(node) * (boxSize.width + runSpacing),
+            (subNodes.length > 1 ? boxSize.height / 2 + spacing / 2 : 0),
           );
       // return 0;
       returnValue = (subNodes.length > 1
           ? boxSize.height * 2 +
-              spacing.dy *
+              spacing *
                   3 // multiplier used to change the distance between the group of nodes
-          : boxSize.height + spacing.dy * 2);
+          : boxSize.height + spacing * 2);
     } else {
       /// if not all are leaves then draw subnodes horizontally
       double dyOff = 0;
@@ -138,13 +161,13 @@ class OrgChartController<E> {
 
       node.position = subNodes.length == 1
           ? Offset(
-              _getLevel(node) * (boxSize.width + spacing.dx),
+              _getLevel(node) * (boxSize.width + runSpacing),
               subNodes.first.position.dy,
             )
           : offset +
               Offset(
-                _getLevel(node) * (boxSize.width + spacing.dx),
-                relOff / 2 - boxSize.height / 2 - spacing.dy,
+                _getLevel(node) * (boxSize.width + runSpacing),
+                relOff / 2 - boxSize.height / 2 - spacing,
               );
       returnValue = relOff;
     }
@@ -158,7 +181,23 @@ class OrgChartController<E> {
   }
 
   /// to remove an item from the list
-  void removeItem(id) {
+  void removeItem(id, ActionOnNodeRemoval action) {
+    assert(toSetter != null,
+        "toSetter is not provided, you can't use this function without providing an toSetter");
+    List subnodes =
+        _nodes.where((element) => toProvider(element.data) == id).toList();
+    switch (action) {
+      case ActionOnNodeRemoval.unlink:
+        for (Node<E> node in subnodes) {
+          toSetter!(node.data, null);
+        }
+        break;
+      case ActionOnNodeRemoval.connectToParent:
+        for (Node<E> node in subnodes) {
+          toSetter!(node.data, toProvider(node.data));
+        }
+        break;
+    }
     _nodes.removeWhere((element) => idProvider(element.data) == id);
     calculatePosition();
   }
@@ -202,10 +241,12 @@ class OrgChartController<E> {
   /// returns the list of root nodes
   List<Node<E>> get roots {
     return _nodes
-        .where((node) => _nodes
-            .where(
-                (element) => idProvider(element.data) == toProvider(node.data))
-            .isEmpty)
+        .where((node) => _getLevel(node) == 1
+            //  _nodes
+            //     .where(
+            //         (element) => idProvider(element.data) == toProvider(node.data))
+            //     .isEmpty
+            )
         .toList();
   }
 
@@ -216,9 +257,9 @@ class OrgChartController<E> {
     _nodes.insert(index == -1 ? math.max(_nodes.length - 1, 0) : index, node);
   }
 
-  // reutns the relative X offset of the node
+  // returns the relative X offset of the node
   double _getRelOffset(Node<E> node) {
-    switch (orientation) {
+    switch (_orientation) {
       case OrgChartOrientation.topToBottom:
         return _getRelOffsetTopToBottom(node);
       case OrgChartOrientation.leftToRight:
@@ -230,27 +271,19 @@ class OrgChartController<E> {
 
   double _getRelOffsetTopToBottom(Node<E> node) {
     List<Node<E>> subNodes = getSubNodes(node);
-    // bool streak = _oneChildStreak(node);
-
     if (node.hideNodes || subNodes.isEmpty) {
-      return boxSize.width + spacing.dx * 2;
+      return boxSize.width + spacing * 2;
     }
 
     double relativeOffset = 0.0;
 
     if (allLeaf(subNodes)) {
       return (subNodes.length > 1
-          ? boxSize.width * 2 + spacing.dx * 3
-          : boxSize.width + spacing.dx * 2);
+          ? boxSize.width * 2 + spacing * 3
+          : boxSize.width + spacing * 2);
     } else {
-      // if (streak) {
-      //   return boxSize.width + spacing.dx;
-      // }
-      // else {
-      // if (_allOnes(node)) return 0;
       for (var i = 0; i < subNodes.length; i++) {
         relativeOffset += _getRelOffsetTopToBottom(subNodes[i]);
-        // }
       }
     }
     return relativeOffset;
@@ -260,17 +293,15 @@ class OrgChartController<E> {
     List<Node<E>> subNodes = getSubNodes(node);
 
     if (node.hideNodes || subNodes.isEmpty) {
-      return boxSize.height + spacing.dy * 2;
+      return boxSize.height + spacing * 2;
     }
 
     double relativeOffset = 0.0;
     if (allLeaf(subNodes)) {
       return (subNodes.length > 1
-          ? boxSize.height * 2 + spacing.dy * 3
-          : boxSize.height + spacing.dy * 2);
+          ? boxSize.height * 2 + spacing * 3
+          : boxSize.height + spacing * 2);
     } else {
-      // if (_allOnes(node)) return 0;
-
       for (var i = 0; i < subNodes.length; i++) {
         relativeOffset += _getRelOffsetLeftToRight(subNodes[i]);
       }
@@ -294,7 +325,7 @@ class OrgChartController<E> {
   /// function recursively calculates the position of the node and its subnodes
   /// returns relative offset of the node
   double _calculateNP(Node<E> node, {Offset offset = const Offset(0, 0)}) {
-    switch (orientation) {
+    switch (_orientation) {
       case OrgChartOrientation.topToBottom:
         return _calculateNPTopToBottom(node, offset: offset);
       case OrgChartOrientation.leftToRight:
@@ -309,9 +340,14 @@ class OrgChartController<E> {
   /// but don't forget to setState after calcutions
   /// this function is called automatically when you change the items list
   void calculatePosition() {
-    Node<E>? n = _nodes.where((element) => _getLevel(element) == 1).firstOrNull;
-    if (n != null) {
-      _calculateNP(n);
+    double offset = 0;
+    for (Node<E> node in roots) {
+      offset += _calculateNP(
+        node,
+        offset: _orientation == OrgChartOrientation.topToBottom
+            ? Offset(offset, 0)
+            : Offset(0, offset),
+      );
       setState?.call(() {});
     }
   }